==> Notebooks/04_Predicting_completion_time.ipynb <==
import joblib
from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LinearRegression




full_pipeline = joblib.load("feature_pipeline_unfitted.pkl")

X = df.drop(columns=["completion_time"])
y = df["completion_time"]

model_pipeline = Pipeline([
    ("features", full_pipeline),
    ("regressor", LinearRegression())
])

scores = cross_val_score(model_pipeline, X, y, cv=5, scoring="neg_mean_squared_error")
print(-scores)
